Major change to project folder structure. Backend program files now reside in  -/Frugality/backend/app
Docker files & main.py/process.py rewritten to adapt. Created a local setup bash script for linux(ubuntu), which creates a virtual env & installs dependencies. Of note: the database url still needs to be manually changed if the backend is run outside of a container (This will be fixed next up) Portforwarding from the container currently also seems to conflict with something on linux (localhost port being set to 8081 instead of 8080 by vscode).